запустил микросервис user-service. Рефакторинг event, compilation, request для работы с ним

+++ core/main-service/src/main/java/ru/practicum/main/service/request/repository/RequestRepository.java
@@ -9,23 +9,23 @@ import ru.practicum.main.service.request.model.RequestStatus;
 import java.util.List;
 import java.util.Optional;
 
-public interface RequestRepository extends JpaRepository<ParticipationRequest, Integer> {
+public interface RequestRepository extends JpaRepository<ParticipationRequest, Long> {
 
-    List<ParticipationRequest> findAllByRequesterId(Integer userId);
+    List<ParticipationRequest> findAllByRequesterId(Long userId);
 
-    Optional<ParticipationRequest> findByIdAndRequesterId(Integer requestId, Integer userId);
+    Optional<ParticipationRequest> findByIdAndRequesterId(Long requestId, Long userId);
 
-    Optional<ParticipationRequest> findByEventIdAndRequesterId(Integer eventId, Integer userId);
+    Optional<ParticipationRequest> findByEventIdAndRequesterId(Long eventId, Long userId);
 
-    List<ParticipationRequest> findAllByEventId(Integer eventId);
+    List<ParticipationRequest> findAllByEventId(Long eventId);
 
-    List<ParticipationRequest> findAllByIdIn(List<Integer> ids);
+    List<ParticipationRequest> findAllByIdIn(List<Long> ids);
 
-    boolean existsByEventIdAndRequesterIdAndStatus(Integer eventId, Integer requesterId, RequestStatus status);
+    boolean existsByEventIdAndRequesterIdAndStatus(Long eventId, Long requesterId, RequestStatus status);
 
     @Query("SELECT COUNT(r) FROM ParticipationRequest r " +
             "WHERE r.event.id = :eventId AND r.status = :status")
-    long countByEventIdAndStatus(@Param("eventId") Integer eventId,
+    long countByEventIdAndStatus(@Param("eventId") Long eventId,
                                  @Param("status") RequestStatus status);
 
     List<ParticipationRequest> findAllByEventIdInAndStatus(List<Long> eventIds, RequestStatus status);
